+++ src/containerapp/azext_containerapp/_help.py
@@ -86,6 +86,98 @@ helps['containerapp list'] = """
           az containerapp list -g MyResourceGroup
 """
 
+helps['containerapp exec'] = """
+    type: command
+    short-summary: Open an SSH-like interactive shell within a container app replica
+    examples:
+    - name: exec into a container app
+      text: |
+          az containerapp exec -n MyContainerapp -g MyResourceGroup
+    - name: exec into a particular container app replica and revision
+      text: |
+          az containerapp exec -n MyContainerapp -g MyResourceGroup --replica MyReplica --revision MyRevision
+    - name: open a bash shell in a containerapp
+      text: |
+          az containerapp exec -n MyContainerapp -g MyResourceGroup --command bash
+"""
+
+helps['containerapp browse'] = """
+    type: command
+    short-summary: Open a containerapp in the browser, if possible
+    examples:
+    - name: open a containerapp in the browser
+      text: |
+          az containerapp browse -n MyContainerapp -g MyResourceGroup
+"""
+
+helps['containerapp up'] = """
+    type: command
+    short-summary: Create or update a container app as well as any associated resources (ACR, resource group, container apps environment, Github Actions, etc.)
+    examples:
+    - name: Create a container app from a dockerfile in a Github repo (setting up github actions)
+      text: |
+          az containerapp up -n MyContainerapp --repo https://github.com/myAccount/myRepo
+    - name: Create a container app from a dockerfile in a local directory
+      text: |
+          az containerapp up -n MyContainerapp --source .
+    - name: Create a container app from an image in a registry
+      text: |
+          az containerapp up -n MyContainerapp --image myregistry.azurecr.io/myImage:myTag
+    - name: Create a container app from an image in a registry with ingress enabled and a specified environment
+      text: |
+          az containerapp up -n MyContainerapp --image myregistry.azurecr.io/myImage:myTag --ingress external --target-port 80 --environment MyEnv
+"""
+
+helps['containerapp logs'] = """
+    type: group
+    short-summary: Show container app logs
+"""
+
+helps['containerapp logs show'] = """
+    type: command
+    short-summary: Show past logs and/or print logs in real time (with the --follow parameter). Note that the logs are only taken from one revision, replica, and container.
+    examples:
+    - name: Fetch the past 20 lines of logs from an app and return
+      text: |
+          az containerapp logs show -n MyContainerapp -g MyResourceGroup
+    - name: Fetch 30 lines of past logs logs from an app and print logs as they come in
+      text: |
+          az containerapp logs show -n MyContainerapp -g MyResourceGroup --follow --tail 30
+    - name: Fetch logs for a particular revision, replica, and container
+      text: |
+          az containerapp logs show -n MyContainerapp -g MyResourceGroup --replica MyReplica --revision MyRevision --container MyContainer
+"""
+
+# Replica Commands
+helps['containerapp replica'] = """
+    type: group
+    short-summary: Manage container app replicas
+"""
+
+helps['containerapp replica list'] = """
+    type: command
+    short-summary: List a container app revision's replica
+    examples:
+    - name: List a container app's replicas in the latest revision
+      text: |
+          az containerapp replica list -n MyContainerapp -g MyResourceGroup
+    - name: List a container app's replicas in a particular revision
+      text: |
+          az containerapp replica list -n MyContainerapp -g MyResourceGroup --revision MyRevision
+"""
+
+helps['containerapp replica show'] = """
+    type: command
+    short-summary: Show a container app replica
+    examples:
+    - name: Show a replica from the latest revision
+      text: |
+          az containerapp replica show -n MyContainerapp -g MyResourceGroup --replica MyReplica
+    - name: Show a replica from the a particular revision
+      text: |
+          az containerapp replica show -n MyContainerapp -g MyResourceGroup --replica MyReplica --revision MyRevision
+"""
+
 # Revision Commands
 helps['containerapp revision'] = """
     type: group
@@ -162,6 +254,14 @@ helps['containerapp revision copy'] = """
 
 """
 
+helps['containerapp revision copy'] = """
+    type: command
+    short-summary: Create a revision based on a previous revision.
+    examples:
+    - name: Create a revision based on a previous revision.
+      text: |
+          az containerapp revision copy -n MyContainerapp -g MyResourceGroup --cpu 0.75 --memory 1.5Gi
+"""
 
 # Environment Commands
 helps['containerapp env'] = """
@@ -176,13 +276,13 @@ helps['containerapp env create'] = """
     - name: Create an environment with an auto-generated Log Analytics workspace.
       text: |
           az containerapp env create -n MyContainerappEnvironment -g MyResourceGroup \\
-              --location "Canada Central"
+              --location eastus2
     - name: Create an environment with an existing Log Analytics workspace.
       text: |
           az containerapp env create -n MyContainerappEnvironment -g MyResourceGroup \\
               --logs-workspace-id myLogsWorkspaceID \\
               --logs-workspace-key myLogsWorkspaceKey \\
-              --location "Canada Central"
+              --location eastus2
 """
 
 
@@ -254,6 +354,61 @@ helps['containerapp env dapr-component remove'] = """
     - name: Remove a Dapr component from a Container Apps environment.
       text: |
           az containerapp env dapr-component remove -g MyResourceGroup --dapr-component-name MyDaprComponentName --name MyEnvironment
+"""
+
+# Identity Commands
+helps['containerapp identity'] = """
+    type: group
+    short-summary: Commands to manage managed identities.
+"""
+
+helps['containerapp identity assign'] = """
+    type: command
+    short-summary: Assign managed identity to a container app.
+    long-summary: Managed identities can be user-assigned or system-assigned.
+    examples:
+    - name: Assign system identity.
+      text: |
+          az containerapp identity assign -n myContainerapp -g MyResourceGroup --system-assigned
+    - name: Assign user identity.
+      text: |
+          az containerapp identity assign -n myContainerapp -g MyResourceGroup --user-assigned myUserIdentityName
+    - name: Assign user identity (from a different resource group than the containerapp).
+      text: |
+          az containerapp identity assign -n myContainerapp -g MyResourceGroup --user-assigned myUserIdentityResourceId
+    - name: Assign system and user identity.
+      text: |
+          az containerapp identity assign -n myContainerapp -g MyResourceGroup --system-assigned --user-assigned myUserIdentityResourceId
+"""
+
+helps['containerapp identity remove'] = """
+    type: command
+    short-summary: Remove a managed identity from a container app.
+    examples:
+    - name: Remove system identity.
+      text: |
+          az containerapp identity remove -n myContainerapp -g MyResourceGroup --system-assigned
+    - name: Remove user identity.
+      text: |
+          az containerapp identity remove -n myContainerapp -g MyResourceGroup --user-assigned myUserIdentityName
+    - name: Remove system and user identity (from a different resource group than the containerapp).
+      text: |
+          az containerapp identity remove -n myContainerapp -g MyResourceGroup --system-assigned --user-assigned myUserIdentityResourceId
+    - name: Remove all user identities.
+      text: |
+          az containerapp identity remove -n myContainerapp -g MyResourceGroup --user-assigned
+    - name: Remove system identity and all user identities.
+      text: |
+          az containerapp identity remove -n myContainerapp -g MyResourceGroup --system-assigned --user-assigned
+"""
+
+helps['containerapp identity show'] = """
+    type: command
+    short-summary: Show managed identities of a container app.
+    examples:
+    - name: Show managed identities.
+      text: |
+          az containerapp identity show -n myContainerapp -g MyResourceGroup
 """
 
 # Ingress Commands
@@ -470,7 +625,7 @@ helps['containerapp dapr'] = """
 
 helps['containerapp dapr enable'] = """
     type: command
-    short-summary: Enable Dapr for a container app.
+    short-summary: Enable Dapr for a container app. Updates existing values.
     examples:
     - name: Enable Dapr for a container app.
       text: |
@@ -479,7 +634,7 @@ helps['containerapp dapr enable'] = """
 
 helps['containerapp dapr disable'] = """
     type: command
-    short-summary: Disable Dapr for a container app.
+    short-summary: Disable Dapr for a container app. Removes existing values.
     examples:
     - name: Disable Dapr for a container app.
       text: |